requirements.txt
anyio==4.9.0
certifi==2025.7.14
click==8.2.1
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pip==24.2
python-telegram-bot==22.3
pywizlight==0.6.3
PyYAML==6.0.2
setuptools==74.1.2
sniffio==1.3.1
tornado==6.5.1
typing_extensions==4.14.1
wheel==0.44.0
